+++ src/librairie/album.h
@@ -6,7 +6,7 @@
 class Album : public Livre
 {
 public:
-    Album(const QString& auteur, const QString& titre, const QString& editeur, int isbn, const QString& public_cible,const QString& illustrations,int code_bibliotheque_origine=-1);
+    Album(const QString& auteur, const QString& titre, const QString& editeur, long long isbn, const QString& public_cible,const QString& illustrations,int code_bibliotheque_origine=-1);
 
     QString getIllustrations();
